feat : upgrade to spring cloud 2024.0.0.0 and localstack to 4.0.3
