+++ aws-kinesis-project/producer/docker/docker-compose.yml
@@ -2,7 +2,7 @@ version: '3.8'
 services:
 
   localstack:
-    image: localstack/localstack:3.8.1
+    image: localstack/localstack:4.0.3
     ports:
       - "4566:4566"
     environment:
